requirements.txt
langchain==0.3.25
langchain-community==0.3.23
langchain-huggingface==0.1.2
chromadb==0.6.3
sentence-transformers==3.4.1
pdfplumber==0.11.4
streamlit==1.45.0
google-generativeai==0.8.5
python-dotenv==1.1.0
onnxruntime